memcg: close the race between migration and installing bprm->mm as mm
Oleg pointed out that there is a race at exec time between when bprm->mm
is initialized and the exec'ing task being migrated to a different memory
control group.

Ractor the code in memcontrol so exec_mmap can use the same code as as
fork to ensure that task->memcg == task->mm->memcg.

Link: http://lkml.kernel.org/r/87fu37cow4.fsf_-_@xmission.com
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Reported-by: Oleg Nesterov <oleg@redhat.com>
Cc: Michal Hocko <mhocko@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Mark Brown <broonie@kernel.org>